add SHARNESS_TEST_NAME and use it in place of SHARNESS_TEST_FILE
SHARNESS_TEST_FILE was used in a couple places in the creation
of new file names, and thus had to be sanitized to remove extension
or path prefix multiple times. Simplify this use by creating a
SHARNESS_TEST_NAME once, and using that where SHARNESS_TEST_FILE
was previously used.