Add SHARNESS_TEST_SRCDIR as path to source of sharness
If tests are run from a different directory than the sources, e.g.

 /path/to/test.t

Then SHARNESS_TEST_DIRECTORY (the current working directory) and
the source of the tests and sharness.sh itself are in different paths.
Differentiate these paths with SHARNESS_TEST_SRCDIR and use this
to load sharness.sh in subdir tests.

This will also be useful when running tests from $srcdir in a $builddir
of a VPATH build of some project.